+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Huff Neo Compiler changelog
 
 ## [Unreleased]
+
+## [1.1.8] - 2025-05-29
 - Convert context to a list of contexts to allow for nested contexts in lexing.
 - Make Prague the default EVM version.
 - Update `revm` and `foundry` to the latest version.